[T155] modified composable function and renamed it to be reusable
