+++ app/src/main/java/com/gathering/android/common/Composables.kt
@@ -30,6 +30,7 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
@@ -52,6 +53,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.rememberAsyncImagePainter
@@ -267,7 +269,7 @@ fun ShowText(
 }
 
 @Composable
-fun ShowImage(imageUri: String?) {
+fun ImageView(imageUri: String?, size: Dp, onClick: () -> Unit) {
     Card(
         modifier = Modifier
             .padding(25.dp)
@@ -280,21 +282,22 @@ fun ShowImage(imageUri: String?) {
         val painter = if (imageUri.isNullOrEmpty()) {
             painterResource(id = R.drawable.ic_person)
         } else {
-            rememberAsyncImagePainter(model = "https://moyar.dev:8080/photo/${imageUri}")
+            rememberAsyncImagePainter(model = imageUri)
         }
         Image(
             contentScale = ContentScale.Crop,
             painter = painter,
-            contentDescription = "Profile picture",
+            contentDescription = "",
             modifier = Modifier
-                .size(200.dp)
+                .size(size)
                 .background(Color.White, CircleShape)
+                .clickable { onClick() }
         )
     }
 }
 
 @Composable
-fun ShowImage(bmp: Bitmap?) {
+fun ImageView(bmp: Bitmap?, size: Dp) {
     Card(
         modifier = Modifier
             .padding(25.dp)
@@ -314,8 +317,19 @@ fun ShowImage(bmp: Bitmap?) {
             painter = painter,
             contentDescription = "",
             modifier = Modifier
-                .size(200.dp)
+                .size(size)
                 .background(Color.White, CircleShape)
         )
     }
+}
+
+@Composable
+fun HorizontalDivider() {
+    Divider(
+        color = Color.LightGray,
+        thickness = 1.dp,
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(1.dp)
+    )
 }
